multimodal_search_demo/requirements.txt
pymilvus>=2.3.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0 